// src/api/openapi/model/offerElementListElement.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OfferElementType } from './offerElementType';


export interface OfferElementListElement { 
    elementType: OfferElementType;
    id: number;
    name: string;
}